Fase 1 is essentially complete. Here's a summary of everything that was done this session:
  StaysPage.tsx — added transfer and addService to destructuring; passed onTransfer and onAddService to StayDrawer.

  StayDrawer.tsx — added onAddService prop, useExtraServices hook, extra services UI section (shows list of applied services + inline form with service selector +
  quantity input).

  CompaniesPage.tsx + CompanyModal.tsx — full CRUD page for companies (mirrors GuestsPage pattern).

  useCompanies.ts — added isUpdating and isDeleting flags.

  Sidebar.tsx / router/index.tsx — wired /companies route and "Empresas" nav entry.

  Backend — ExtraServiceController (GET /extra-services), StayController::addService (POST /stays/{id}/services), ExtraServicesSeeder with 8 default services, routes
  registered.

  Types — ExtraService, StayService interfaces; services? on Stay.

+++ backend/app/Http/Controllers/Api/V1/AuthController.php
@@ -36,7 +36,7 @@ class AuthController extends Controller
         $user->tokens()->delete();
 
         $token = $user->createToken('hotel-session', ['*'], now()->addMinutes(
-            config('sanctum.expiration', 480)
+            (int) config('sanctum.expiration', 480)
         ))->plainTextToken;
 
         ActivityLog::record('login', $user->id, ['ip' => $request->ip()]);
